group_3_gene_path_ai/requirements.txt
pydeseq2
pandas
numpy
gseapy
streamlit
huggingface_hub
python-dotenv
psutil
plotly
mygene
polars